separated cmds from index.js into individ. files
